Changed int to long
Changed any references of int to long so that the sequence could accept arbitrarily large numbers.

+++ src/api/ComputeEngineAPI.java
@@ -11,5 +11,5 @@ public interface ComputeEngineAPI {
 	ComputeCompleteResponse completeCompute(ComputeCompleteRequest request);
 
 	// Collatz sequence as a string
-	String collatzSequenceString(int n);
+	String collatzSequenceString(long n);
 }
